+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 1.0.20
+
+* Include uthash.h in hex package
+
+# Version 1.0.19
+
+* Fix couple memory leaks
+* Switch hashmap to uthash library
+* Use system allocator in openssl
+* Update ciphers and option to safer defaults
+
 # Version 1.0.18
 
 * Do not report badly formed Client Hello as a TLS error